Make impl IntoDeserializer for &[u8] public
